Cambio tipo de timestamp de double a uint64_t. (está así en el enunciado del tp)

+++ Pool-Memorias/memoria.c
@@ -1184,7 +1184,7 @@ int pasarValoresALisandra(datosJournal* datos,int socket_lfs)
 	char aux[100];
 	int cont = 0;
 	while(enviar != NULL){
-		snprintf(aux,100,"INSERT %s %d %s %ld",enviar->nombreTabla,enviar->key,enviar->value,(uint64_t) enviar->timestamp);
+		snprintf(aux,100,"INSERT %s %d %s %ld",enviar->nombreTabla,enviar->key,enviar->value, enviar->timestamp);
 		insert.tam = strlen(aux)+1;
 		insert.str = malloc(insert.tam);
 		strcpy(insert.str,aux);
